requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
sentence-transformers
PyPDF2
python-multipart
numpy
requests
mysql-connector-python